Moved Content to wwwroot, updated references, fixed standalone radio page, include razor content in nuget, version bump
