+++ GovUk.Frontend.AspNetCore.Extensions/GovUk.Frontend.AspNetCore.Extensions.csproj
@@ -4,21 +4,27 @@
 		<TargetFramework>net6.0</TargetFramework>
 		<AddRazorSupportForMvc>true</AddRazorSupportForMvc>
 		<GenerateEmbeddedFilesManifest>true</GenerateEmbeddedFilesManifest>
+		<IncludeRazorContentInPack>true</IncludeRazorContentInPack>
 		<Nullable>enable</Nullable>
-		<Version>2.0.0</Version>
+		<Version>2.0.1</Version>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<None Remove="Content\govuk\govuk-frontend.css" />
-		<None Remove="Content\govuk\govuk-js-init.js" />
-		<None Remove="Content\govuk\govuk-js-test.js" />
-		<None Remove="Content\govuk\govuk-validation.js" />
-		<None Remove="Content\govuk\open-sans-bold.woff2" />
-		<None Remove="Content\govuk\open-sans-regular.woff2" />
-		<None Remove="Content\govuk\open-sans-semi-bold.woff2" />
-		<None Remove="Content\tpr\chevron_right_eclipse_24dp.svg" />
-		<None Remove="Content\tpr\chevron_right_white_24dp.svg" />
-		<None Remove="Content\tpr\tpr.css" />
+	  <Content Remove="wwwroot\govuk\govuk-frontend.css" />
+	  <Content Remove="wwwroot\govuk\govuk-frontend.css.map" />
+	  <Content Remove="wwwroot\govuk\govuk-js-init.js" />
+	  <Content Remove="wwwroot\govuk\govuk-js-test.js" />
+	  <Content Remove="wwwroot\govuk\govuk-validation.js" />
+	  <Content Remove="wwwroot\govuk\open-sans-bold.woff2" />
+	  <Content Remove="wwwroot\govuk\open-sans-regular.woff2" />
+	  <Content Remove="wwwroot\govuk\open-sans-semi-bold.woff2" />
+	  <Content Remove="wwwroot\tpr\chevron_right_eclipse_24dp.svg" />
+	  <Content Remove="wwwroot\tpr\chevron_right_white_24dp.svg" />
+	  <Content Remove="wwwroot\tpr\tpr.css" />
+	  <Content Remove="wwwroot\tpr\tpr.css.map" />
+	</ItemGroup>
+
+	<ItemGroup>
 		<None Remove="Styles\_govuk-button.scss" />
 		<None Remove="Styles\_govuk-caption.scss" />
 		<None Remove="Styles\_govuk-date-input.scss" />
@@ -28,20 +34,13 @@
 		<None Remove="Styles\_govuk-radios-checkboxes.scss" />
 		<None Remove="Styles\_tpr-grid.scss" />
 		<None Remove="Styles\_variables.scss" />
-		<None Remove="Content\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.min.js" />
-		<None Remove="Content\lib\jquery-validation\dist\jquery.validate.min.js" />
-		<None Remove="Content\lib\jquery\dist\jquery.min.js" />
-		<None Remove="Content\lib\jquery\dist\jquery.min.map" />
+		<None Remove="wwwroot\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.min.js" />
+		<None Remove="wwwroot\lib\jquery-validation\dist\jquery.validate.min.js" />
+		<None Remove="wwwroot\lib\jquery\dist\jquery.min.js" />
+		<None Remove="wwwroot\lib\jquery\dist\jquery.min.map" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<EmbeddedResource Include="Content\govuk\govuk-frontend.css" />
-		<EmbeddedResource Include="Content\govuk\open-sans-bold.woff2" />
-		<EmbeddedResource Include="Content\govuk\open-sans-regular.woff2" />
-		<EmbeddedResource Include="Content\govuk\open-sans-semi-bold.woff2" />
-		<EmbeddedResource Include="Content\tpr\chevron_right_white_24dp.svg" />
-		<EmbeddedResource Include="Content\tpr\chevron_right_eclipse_24dp.svg" />
-		<EmbeddedResource Include="Content\tpr\tpr.css" />
 		<EmbeddedResource Include="Styles\_govuk-caption.scss" />
 		<EmbeddedResource Include="Styles\_govuk-button.scss" />
 		<EmbeddedResource Include="Styles\_govuk-date-input.scss" />
@@ -52,13 +51,20 @@
 		<EmbeddedResource Include="Styles\_govuk-link.scss" />
 		<EmbeddedResource Include="Styles\_variables.scss" />
 		<EmbeddedResource Include="Styles\_tpr-back-to-top.scss" />
-		<EmbeddedResource Include="Content\govuk\govuk-js-init.js" />
-		<EmbeddedResource Include="Content\govuk\govuk-js-test.js" />
-		<EmbeddedResource Include="Content\govuk\govuk-validation.js" />
-		<EmbeddedResource Include="Content\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.min.js" />
-		<EmbeddedResource Include="Content\lib\jquery-validation\dist\jquery.validate.min.js" />
-		<EmbeddedResource Include="Content\lib\jquery\dist\jquery.min.js" />
-		<EmbeddedResource Include="Content\lib\jquery\dist\jquery.min.map" />
+		<EmbeddedResource Include="wwwroot\govuk\govuk-frontend.css" />
+		<EmbeddedResource Include="wwwroot\govuk\govuk-js-init.js" />
+		<EmbeddedResource Include="wwwroot\govuk\govuk-js-test.js" />
+		<EmbeddedResource Include="wwwroot\govuk\govuk-validation.js" />
+		<EmbeddedResource Include="wwwroot\govuk\open-sans-bold.woff2" />
+		<EmbeddedResource Include="wwwroot\govuk\open-sans-regular.woff2" />
+		<EmbeddedResource Include="wwwroot\govuk\open-sans-semi-bold.woff2" />
+		<EmbeddedResource Include="wwwroot\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.min.js" />
+		<EmbeddedResource Include="wwwroot\lib\jquery-validation\dist\jquery.validate.min.js" />
+		<EmbeddedResource Include="wwwroot\lib\jquery\dist\jquery.min.js" />
+		<EmbeddedResource Include="wwwroot\lib\jquery\dist\jquery.min.map" />
+		<EmbeddedResource Include="wwwroot\tpr\chevron_right_eclipse_24dp.svg" />
+		<EmbeddedResource Include="wwwroot\tpr\chevron_right_white_24dp.svg" />
+		<EmbeddedResource Include="wwwroot\tpr\tpr.css" />
 	</ItemGroup>
 
 	<ItemGroup>
@@ -183,6 +189,23 @@
 	
 	<ItemGroup>
 	  <UpToDateCheckInput Remove="Styles\_tpr-add-link.scss" />
+	</ItemGroup>
+	
+	<ItemGroup>
+	  <None Include="wwwroot\govuk\govuk-frontend.css.map" />
+	  <None Include="wwwroot\lib\jquery-validation-unobtrusive\jquery.validate.unobtrusive.js" />
+	  <None Include="wwwroot\lib\jquery-validation-unobtrusive\LICENSE.txt" />
+	  <None Include="wwwroot\lib\jquery-validation\dist\additional-methods.js" />
+	  <None Include="wwwroot\lib\jquery-validation\dist\additional-methods.min.js" />
+	  <None Include="wwwroot\lib\jquery-validation\dist\jquery.validate.js" />
+	  <None Include="wwwroot\lib\jquery-validation\LICENSE.md" />
+	  <None Include="wwwroot\lib\jquery\dist\jquery.js" />
+	  <None Include="wwwroot\lib\jquery\LICENSE.txt" />
+	  <None Include="wwwroot\tpr\tpr.css.map" />
+	</ItemGroup>
+	
+	<ItemGroup>
+	  <Folder Include="Content\" />
 	</ItemGroup>
 
 </Project>
