+++ GovUk.Frontend.AspNetCore.Extensions/__tests__/govuk-validation-validateElement.tests.js
@@ -1,4 +1,4 @@
-const govuk = require("../Content/govuk/govuk-validation");
+const govuk = require("../wwwroot/govuk/govuk-validation");
 
 let _mockUpdateError;
 const _mockValidator = {
