+++ GovUk.Frontend.AspNetCore.Extensions/__tests__/govuk-validation-validatePhone.tests.js
@@ -1,4 +1,4 @@
-const govuk = require("../Content/govuk/govuk-validation");
+const govuk = require("../wwwroot/govuk/govuk-validation");
 
 describe("validatePhone", () => {
   it("should allow valid patterns", () => {
