+++ GovUk.Frontend.AspNetCore.Extensions/wwwroot/govuk/govuk-js-init.js
