+++ GovUk.Frontend.ExampleApp/Views/Radios/Index.cshtml
@@ -11,25 +11,28 @@
 <form asp-controller="Radios" asp-action="Post" method="post" novalidate>
     <govuk-client-side-validation>
         <govuk-radios asp-for="Field1">
-        <govuk-radios-fieldset>
-            <govuk-radios-fieldset-legend>This is the legend</govuk-radios-fieldset-legend>
-        </govuk-radios-fieldset>
-        <govuk-radios-hint>This is the hint</govuk-radios-hint>
-        <govuk-radios-item value="1">Item one (reveals related question)
-            <govuk-radios-item-conditional>
-                <govuk-input asp-for="Field2"><govuk-input-label>Answer this related question</govuk-input-label></govuk-input>
-            </govuk-radios-item-conditional>
-        </govuk-radios-item>
-        <govuk-radios-item value="2">Item two
-            <govuk-radios-item-hint>Optional hint text</govuk-radios-item-hint>
-        </govuk-radios-item>
-        <govuk-radios-divider>or</govuk-radios-divider>
-        <govuk-radios-item value="3">Item three</govuk-radios-item>
+            <govuk-radios-fieldset>
+                <govuk-radios-fieldset-legend>This is the legend</govuk-radios-fieldset-legend>
+                <govuk-radios-hint>This is the hint</govuk-radios-hint>
+                <govuk-radios-item value="1">
+                    Item one (reveals related question)
+                    <govuk-radios-item-conditional>
+                        <govuk-input asp-for="Field2"><govuk-input-label>Answer this related question</govuk-input-label></govuk-input>
+                    </govuk-radios-item-conditional>
+                </govuk-radios-item>
+                <govuk-radios-item value="2">
+                    Item two
+                    <govuk-radios-item-hint>Optional hint text</govuk-radios-item-hint>
+                </govuk-radios-item>
+                <govuk-radios-divider>or</govuk-radios-divider>
+                <govuk-radios-item value="3">Item three</govuk-radios-item>
+            </govuk-radios-fieldset>
+
         </govuk-radios>
     </govuk-client-side-validation>
 
     <govuk-button>Submit</govuk-button>
 </form>
 @section scripts {
-    <partial name="GOVUK/Validation" />
+<partial name="GOVUK/Validation" />
 }
